orchestrator-4/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
minio==7.2.7
